feat(v0.0.9): audit remediation — Decimal correctness, honest coverage gate, CLI dedupe
Correctness
- Decimal end-to-end for all money fields; garbled or missing amounts
  now raise instead of silently becoming 0.0 (new _amounts.py helpers)
- Dedup keys are occurrence-counted (<hash>:<n>) and include
  transaction_id/reference; fixed index-space bug in
  Deduplicator.deduplicate(); per-file scan dedup in scanner

Security
- CAMT recovery-mode parsing is opt-in (allow_recovery=True); strict
  lxml default rejects billion-laughs/undeclared entities at parse time
- SECURITY.md: new "XML Parsing: Why lxml" section

Architecture
- LLM response parsing consolidated into bankstatementparser/_llm_common.py
  (was duplicated across LLMExtractor, VisionExtractor, categorizer,
  ollama_direct)
- parse_camt/parse_pain delegate to shared
  BankStatementCLI._parse_statement_file; public API unchanged
- bank_statement_parsers wrappers now emit DeprecationWarning

Testing
- Golden-file corpus (tests/test_data/golden/) + test_golden_files.py,
  incl. multi-currency CAMT and German CSV headers
- Honest coverage gate: --cov-fail-under 100 -> 98 (measured 99.48%),
  codecov target:auto fails PRs that regress vs. base
- test_branch_coverage.py rewritten as behavior-framed
  test_parser_edge_behavior.py; test_coverage_gaps.py renamed to
  test_error_paths.py with behavioral docstrings
- 703 tests pass; ruff, mypy --strict, bandit all clean

Docs
- README/FAQ/CONTRIBUTING/compliance docs: "100% branch coverage" and
  "697 tests" claims replaced with real figures (703 tests, 31 modules)

+++ bankstatementparser/record_types.py
@@ -17,11 +17,12 @@
 
 from __future__ import annotations
 
+from decimal import Decimal
 from typing import TypedDict
 
 
 class BalanceRecord(TypedDict, total=False):
-    Amount: float
+    Amount: Decimal
     Currency: str | None
     Code: str | None
     Description: str | None
@@ -31,7 +32,7 @@ class BalanceRecord(TypedDict, total=False):
 
 
 class TransactionRecord(TypedDict, total=False):
-    Amount: float
+    Amount: Decimal
     Currency: str | None
     DrCr: str | None
     Debtor: str | None
@@ -44,7 +45,7 @@ class TransactionRecord(TypedDict, total=False):
     CreditorAddress: str | None
     date: str | None
     description: str | None
-    amount: float | None
+    amount: Decimal | None
     currency: str | None
     balance: object
     account_id: str | None
@@ -78,16 +79,16 @@ class StatementStatsRecord(TypedDict, total=False):
     AccountId: str | None
     StatementCreated: str | None
     NumTransactions: int
-    NetAmount: float
+    NetAmount: Decimal
 
 
 class SummaryRecord(TypedDict, total=False):
     account_id: str | None
     statement_date: str | None
     transaction_count: int
-    total_amount: float
-    opening_balance: float | None
-    closing_balance: float | None
+    total_amount: Decimal
+    opening_balance: Decimal | None
+    closing_balance: Decimal | None
     currency: str | None
     message_id: str | None
     initiating_party: str | None
